separate menu and input tests
