+++ spec/plugin/commandline/menu_spec.rb
@@ -3,7 +3,7 @@
 require 'spec_helper'
 
 describe 'esearch#cmdline menu' do
-  include Helpers::CommandlineMenu
+  include Helpers::Commandline
 
   shared_examples 'commandline menu testing examples' do
     let(:open_menu) { '\\<C-o>' }
@@ -212,177 +212,6 @@ describe 'esearch#cmdline menu' do
 
         include_examples 'it puts cursor at the after dismissing with', keys: [:enter]
         include_examples 'it puts cursor at the after dismissing with', keys: [:escape]
-      end
-    end
-
-    context 'when cancelling initial selection' do
-      before { esearch.configuration.submit!(overwrite: true) } # TODO: will be removed
-
-      shared_examples 'it starts search at location "|" after pressing' do |keys:, prefilled_input:, expected_input:|
-        context "when #{keys} are pressed" do
-          before { expect(keys.size).to be_present, "wrong usage: must be at least 1 char to submit search" }
-
-          it 'it starts search at a specific location' do
-            editor.send_keys(*open_input, prefilled_input, :enter)
-            expect(esearch).to finish_search_for(prefilled_input)
-
-              expect {
-                editor.send_keys(*open_input)
-                next if keys.size < 2
-
-                editor.send_keys_separately(*keys[..-2])
-                expect(editor.commandline_cursor_location).to eq(expected_input.index('|')+1)
-              }.not_to start_search
-
-            expect { editor.send_keys(keys[-1]) }
-              .to start_search
-              .and finish_search_for(expected_input.tr('|', ''))
-          end
-        end
-      end
-
-      shared_examples "it doesn't start search after pressing" do |keys:, prefilled_input:|
-        context "when #{keys} are pressed" do
-          it "it doesn't start search" do
-            editor.send_keys(*open_input, prefilled_input, :enter)
-            expect(esearch).to finish_search_for(prefilled_input)
-
-              expect {
-                editor.send_keys(*open_input)
-                editor.send_keys_separately(*keys)
-              }.not_to start_search
-          end
-        end
-      end
-
-      context 'cancelling prefilled input selection' do
-        context 'with moving cursor' do
-          include_examples 'it starts search at location "|" after pressing',
-            keys:            ['\\<Left>', :enter],
-            prefilled_input: 'str',
-            expected_input:  'st|r'
-
-          include_examples 'it starts search at location "|" after pressing',
-            keys:            ['\\<Right>', :enter],
-            prefilled_input: 'str',
-            expected_input:  'str|'
-
-          include_examples 'it starts search at location "|" after pressing',
-            keys:            ['\\<S-Left>', :enter],
-            prefilled_input: 'str',
-            expected_input:  '|str'
-
-          include_examples 'it starts search at location "|" after pressing',
-            keys:            ['\\<S-Right>', :enter],
-            prefilled_input: 'str',
-            expected_input:  'str|'
-
-          include_examples 'it starts search at location "|" after pressing',
-            keys:            ['\\<C-a>', :enter],
-            prefilled_input: 'str',
-            expected_input:  'str|'
-          include_examples 'it starts search at location "|" after pressing',
-            keys:            ['\\<C-e>', :enter],
-            prefilled_input: 'str',
-            expected_input:  'str|'
-
-          context 'up and down keys' do
-            include_context 'fix vim internal quirks with mapping timeout'
-            include_examples 'it starts search at location "|" after pressing',
-              keys:            %i[up down enter],
-              prefilled_input: 'str',
-              expected_input:  'str|'
-            end
-        end
-
-        context 'without moving cursor' do
-          include_examples 'it starts search at location "|" after pressing',
-            keys:            ['\\<C-c>', :enter],
-            prefilled_input: 'str',
-            expected_input:  'str|'
-          include_examples 'it starts search at location "|" after pressing',
-            keys:            [:escape, :enter],
-            prefilled_input: 'str',
-            expected_input:  'str|'
-        end
-
-        context 'with pressing remapped hotkeys' do
-          context 'defaults' do
-            include_examples 'it starts search at location "|" after pressing',
-              keys:            ['\\<C-o>', :escape, :enter],
-              prefilled_input: 'str',
-              expected_input:  'str|'
-          end
-
-          context 'alt-f' do
-            include_context 'defined commandline hotkey', '<M-f>', '<S-Right>'
-            include_examples 'it starts search at location "|" after pressing',
-              keys:            ['\\<M-f>', :enter],
-              prefilled_input: 'str',
-              expected_input:  'str|'
-          end
-
-          context 'alt-b' do
-            include_context 'defined commandline hotkey', '<M-b>', '<S-Left>'
-            include_examples 'it starts search at location "|" after pressing',
-              keys:            ['\\<M-b>', :enter],
-              prefilled_input: 'str',
-              expected_input:  '|str'
-          end
-        end
-      end
-
-      context 'starting search with prefilled text skipping input step' do
-        context 'when default keys' do
-          include_examples 'it starts search at location "|" after pressing',
-            keys:            [:enter],
-            prefilled_input: 'input was',
-            expected_input:  'input was|'
-        end
-
-        context 'when custom keys' do
-          context 'when defined' do
-            before { editor.command('call add(g:esearch#cmdline#start_search_chars, "s")') }
-            after { editor.command('unlet g:esearch#cmdline#start_search_chars[-1]') }
-
-            include_examples 'it starts search at location "|" after pressing',
-              keys:            ["s"],
-              prefilled_input: 'input was',
-              expected_input:  'input was|'
-          end
-
-          context 'when not defined' do
-            include_examples "it doesn't start search after pressing",
-              keys:            ["s"],
-              prefilled_input: 'input was'
-          end
-        end
-      end
-
-      context 'overriding prefilled input selection' do
-        include_examples 'it starts search at location "|" after pressing',
-          keys:            [:delete, 'str', :enter],
-          prefilled_input: 'input was',
-          expected_input:  'str|'
-
-        include_examples 'it starts search at location "|" after pressing',
-          keys:            [:backspace, 'str', :enter],
-          prefilled_input: 'input was',
-          expected_input:  'str|'
-
-        context 'single char' do
-          include_examples 'it starts search at location "|" after pressing',
-            keys:            ['1', :enter],
-            prefilled_input: 'input was',
-            expected_input:  '1|'
-        end
-
-        context 'multiple chars' do
-          include_examples 'it starts search at location "|" after pressing',
-            keys:            ['multiple chars', :enter],
-            prefilled_input: 'input was',
-            expected_input:  'multiple chars|'
-        end
       end
     end
   end
